--- cert/cert.pem
@@ -1,22 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIDnTCCAoUCFF6rC9ViXLOa+RNLCiuKmbzT56AQMA0GCSqGSIb3DQEBCwUAMIGK
-MQswCQYDVQQGEwJDTjEQMA4GA1UECAwHSmlhbmdzdTEPMA0GA1UEBwwGU3V6aG91
-MQ0wCwYDVQQKDARNU0ZUMQ4wDAYDVQQLDAVUZWFtczEVMBMGA1UEAwwMbHViby1k
-ZXNrdG9wMSIwIAYJKoZIhvcNAQkBFhNib2x1MUBtaWNyb3NvZnQuY29tMB4XDTIx
-MDIxOTEwMzIyMloXDTQ4MDcwNjEwMzIyMlowgYoxCzAJBgNVBAYTAkNOMRAwDgYD
-VQQIDAdKaWFuZ3N1MQ8wDQYDVQQHDAZTdXpob3UxDTALBgNVBAoMBE1TRlQxDjAM
-BgNVBAsMBVRlYW1zMRUwEwYDVQQDDAxsdWJvLWRlc2t0b3AxIjAgBgkqhkiG9w0B
-CQEWE2JvbHUxQG1pY3Jvc29mdC5jb20wggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAw
-ggEKAoIBAQDEto6eKS0wH1mDKO6fy6ZG752Q/i2/FN4/hrFfftpotrB5LrhmP3mU
-Roe2eDfcntEsJt/PZP6FHcEuXxvJ/TuUy4wbKI6hHRq/1NarFN821jnrmClwjFzs
-3TPf0rXW5ENFmDii86vUwjq6tDEq6dvMZmkMAbpDz7g8N6mSNFuAKGtyL7dcDqfA
-IPxOfroASeTFoMy1DXUTkb03bpCltCCIpTpoOekkxJVWMgBxJonihIOpau9qpVdQ
-jLl3Ob4APGKmkIFCWySkk4viFii9udgDuQB2b8f/9AC5don9Rqnzeg4qsW6dOo5v
-PjYS+dytG71zgYCgR8z71YTx/DEK+R5pAgMBAAEwDQYJKoZIhvcNAQELBQADggEB
-ACvRXVb2gKmtiWPz9eJNbEFPbXH/3H/jnSfR21zJfbx/ufksFp/6w4Y+oxf9SVCq
-QVRQRXiopOaLVCO2eqgtG1c6mB5DDYA6WATw1rA1Er+AFjvjbibcIs5hjacOi+iY
-WiFozzHNFf/CfoGLCv5k6ai4czVJzJeVMR8QWNgQHoKacyxkgx+p9RPGLLLvmWQf
-J8zN6TyvcadJTHo7E2wBYq3InBlK9GZPv6/1gXlCPQpgJsrG3lVCS2vrAJwPUPzn
-LqiK98X+4cydZselv3Mzw7fNBSKcmRWvpcHsde3Q3L3516ldBifFm6ad1Y8JqPza
-CmhsQK2XMghKaWsdfG88vPE=
------END CERTIFICATE-----
