--- cert/csr.pem
@@ -1,18 +1,0 @@
------BEGIN CERTIFICATE REQUEST-----
-MIIC5jCCAc4CAQAwgYoxCzAJBgNVBAYTAkNOMRAwDgYDVQQIDAdKaWFuZ3N1MQ8w
-DQYDVQQHDAZTdXpob3UxDTALBgNVBAoMBE1TRlQxDjAMBgNVBAsMBVRlYW1zMRUw
-EwYDVQQDDAxsdWJvLWRlc2t0b3AxIjAgBgkqhkiG9w0BCQEWE2JvbHUxQG1pY3Jv
-c29mdC5jb20wggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQDEto6eKS0w
-H1mDKO6fy6ZG752Q/i2/FN4/hrFfftpotrB5LrhmP3mURoe2eDfcntEsJt/PZP6F
-HcEuXxvJ/TuUy4wbKI6hHRq/1NarFN821jnrmClwjFzs3TPf0rXW5ENFmDii86vU
-wjq6tDEq6dvMZmkMAbpDz7g8N6mSNFuAKGtyL7dcDqfAIPxOfroASeTFoMy1DXUT
-kb03bpCltCCIpTpoOekkxJVWMgBxJonihIOpau9qpVdQjLl3Ob4APGKmkIFCWySk
-k4viFii9udgDuQB2b8f/9AC5don9Rqnzeg4qsW6dOo5vPjYS+dytG71zgYCgR8z7
-1YTx/DEK+R5pAgMBAAGgFjAUBgkqhkiG9w0BCQcxBwwFMTIzNDUwDQYJKoZIhvcN
-AQELBQADggEBABqZveaEbQrmqOnmBw5g7SGl7jqdmkFDPoyWGwQ0fx6sv52aJ3rd
-+FI1Cda7sTnCLyKxWRGCaZl2GdvZ57rGriKZ5qhC70rCHFWls9OzaJDxd9FPwjOR
-YYPbcua4hRvsUuOwj/1S/k5zse1ZMwVrdo+f7i76Nr5v3YVZp0n8exDgDuLwHUuG
-6RbUp8oZ25qW3UlwLEnJU6jpJjRbRXPK+R8Hl760PZipum0Y39/3iR6u4NTh/+AZ
-RNEPBfcHZ8+FcdRP2BwhRZ9ZQz+BXgVwfddLPre7V1iBbrd4a4WIh+tTTv5UfdBz
-wLkmgyMWAezvmftmH7pH/3ptnn9tiZqx6Gs=
------END CERTIFICATE REQUEST-----
