--- cert/key.pem
@@ -1,27 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIIEpQIBAAKCAQEAxLaOniktMB9Zgyjun8umRu+dkP4tvxTeP4axX37aaLaweS64
-Zj95lEaHtng33J7RLCbfz2T+hR3BLl8byf07lMuMGyiOoR0av9TWqxTfNtY565gp
-cIxc7N0z39K11uRDRZg4ovOr1MI6urQxKunbzGZpDAG6Q8+4PDepkjRbgChrci+3
-XA6nwCD8Tn66AEnkxaDMtQ11E5G9N26QpbQgiKU6aDnpJMSVVjIAcSaJ4oSDqWrv
-aqVXUIy5dzm+ADxippCBQlskpJOL4hYovbnYA7kAdm/H//QAuXaJ/Uap83oOKrFu
-nTqObz42EvncrRu9c4GAoEfM+9WE8fwxCvkeaQIDAQABAoIBAHKK3pan4WEhEA04
-e2Jewe6SPgychOS78QHKm2PAyMqqfi3CD6uyEEhPnVyZueVkMOZlqHLqGu5GOiOX
-S+Yc52Gn+GgkAJE+EkJbW7h4ydTzfvx5OMAJa16dG0ScPDfp6GYQ0gHIo2iaxHnM
-zMhxLgRiKY7NPrmXrW6a9slkZVdEqnO3laIgRSOG5N898AF3dPRUlsK3o6Wy0UfR
-YCU62g/7X+GiY4TFpC3j3RKPc2ke+Jqysh5fwPHf63b1X8cUWNnQVZdvuO3IVfhb
-vH8aUeBInUQb9B0uMt6z6tkzJ0+mAEwbhy7kY3ADF5nc5d32mbsbcMGxIICNNFxt
-UCpoR1ECgYEA/y1VrgBogjas9bVD0EZ6lqcnL2EC6qN0DRIN1gqqzgT2rf0AtA8+
-tDBahAt4/zJc3Ca8SmcwdJn4EHE+OcCfGLJcQBMY7PEAEU/OufmQFbIsF9VO1ew2
-HNHQQ/oJvmz5NkxfbJFcZLUOLBdtXbL1qtjvLREGvYArIDoSAxCe+7UCgYEAxVj0
-41R1TrwU2crHoIwz7bopEOKt+3WKOv4b68EPTC0RBba2JNk/KpaJRBjnOkbzbl7R
-Ja/uUAQXN5yUtIM1Jdke+Ql+r1DfkvVkJUv2gArCiDl9JQbTca5ocesgvcFTy+uG
-EDmisoXPvEru0rgifg+sMDLYsIjE5GwgDsJCkGUCgYEArllhUzQANAf7gR5KqUGe
-fNTUlzCnd34aytttDLPZJE6L5TEai+xmoWLwVYwpLeRRQjJNMGD04JZawZE+PON5
-ULx8/2KNA0dfDnUDAFqdsdU8/0PFS7DLAEHzFJ8AjVWTqR/hiyJ+iGiUuFeBaCXg
-aM00Dgy2GEA8Nxa2O9XdaokCgYEAgWF+u4NlnQblzJuLIRCKAvMndTCHy/IVe/OY
-rolV1CU/obVV2Pas2Oj0eTaycZyOf++dws2CuebUqdr4CxhyQawFCFRXP+0PhrNm
-wqT4qJ1lSBYYm6vL+B3Vz8nNH1qd63Uv4XdH2S6tMM8XoAKkam0wADF1qtcf7nmM
-0uA875kCgYEAmyMSIdi/h7HvEp0BBqWvTACExOcb8lNpEmNVUoiJJfjey3+XqnVy
-4K5Tx+K9Hay7age4URiU3bCMKl6fPLah0uIXlR8EMY8KwGe5t7CJ9n33ANGISbjh
-8QInYbhxVmF5DRqag1TZX+e42wlF77CSOPppdJiqxvHqG56J+Vy125s=
------END RSA PRIVATE KEY-----
