+++ index.html
@@ -1,3 +1,18 @@
+<!-- <!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Vite App</title>
+  </head>
+  <body>
+    <div id="app"></div>
+    <script type="module" src="/src/index.js"></script>
+  </body>
+</html> -->
+
+
 <!DOCTYPE html>
 <html lang="en">
 <head>
@@ -16,9 +31,7 @@
         <a class="thumbnail"></a>
       </div>
     </div>
-    <link rel="stylesheet" type="text/css" href="index.css" />
-    <script src="./sdk/dist/MicrosoftTeams.js"></script>
-    <script src="./webgl-video-filter.js"></script>
-    <script src="./index.js"></script>
+    <link rel="stylesheet" type="text/css" href="./src/index.css" />
+    <script src="./src/index.js" type="module"></script>
 </body>
 </html>
