+++ src/index.css
@@ -28,13 +28,13 @@
 #filter-half > a {
   width: 74px;
   height: 42px;
-  background: url(./img/half.png);
+  background: url(/img/half.png);
 }
 
 #filter-gray > a {
     width: 74px;
     height: 42px;
-  background: url(./img/gray.png);
+  background: url(/img/gray.png);
 }
 
 .app-title {
